requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
app-store-scraper==0.3.5
bertopic==0.17.4
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
fastapi==0.135.1
filelock==3.25.2
firebase_admin==7.2.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-play-scraper==1.2.7
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hdbscan==0.8.41
hf-xet==1.4.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.6.0
hyperframe==6.1.0
idna==2.10
Jinja2==3.1.6
joblib==1.5.3
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.18.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
plotly==6.6.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.0
pynndescent==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
umap-learn==0.5.11
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
